Update PRIORITY_UPDATE handling as per RFC9218

+++ lib/common/http3client.c
@@ -306,7 +306,7 @@ static void on_getaddr(h2o_hostinfo_getaddr_req_t *getaddr_req, const char *errs
     start_connect(conn, selected->ai_addr);
 }
 
-static void handle_control_stream_frame(h2o_http3_conn_t *_conn, uint8_t type, const uint8_t *payload, size_t len)
+static void handle_control_stream_frame(h2o_http3_conn_t *_conn, uint64_t type, const uint8_t *payload, size_t len)
 {
     struct st_h2o_httpclient__h3_conn_t *conn = (void *)_conn;
     int err;
